requirements.txt
pipecat-ai[deepgram,openai,silero,runner,local-smart-turn-v3,webrtc]>=0.0.50
python-dotenv>=1.0.0
fastapi>=0.115.6
uvicorn>=0.32.0
aiohttp>=3.11.0
edge-tts>=7.0.0
loguru>=0.7.3
